Moved PlayableCard related patches in the ability classes to own class.
